backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "org-copilot-backend"
version = "0.8.0"
description = "OrgCopilot FastAPI backend (RAG / knowledge base)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy[asyncio]>=2.0.25",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0",
    "httpx>=0.26.0",
    "cryptography>=42.0.0",
    "pymilvus[milvus-lite]>=2.4.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "greenlet>=3.0.0",
    "fastembed>=0.3.0",
    "Pillow>=10.0.0",
    # milvus-lite 仍依赖 pkg_resources；setuptools 81+ 起弃用/移除趋势，锁在 80.x 以兼容 Lite
    "setuptools>=70.0.0,<81",
    "llama-index-core>=0.12.0,<0.13",
    "llama-index-llms-openai>=0.3.0,<0.4",
    "llama-index-embeddings-openai>=0.3.0,<0.4",
    "sqlparse>=0.5.0",
]

[project.optional-dependencies]
# 知识库图片 OCR（PaddleOCR）；安装：uv sync --extra image
image = [
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0,<3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
